Use argparse
Replace the ad-hoc argument parsing code with the argparse[1] parser,
which automatically handles missing required arguments and avoids magic
numbers in referencing positional arguments. Also document/validate that
the environment is one of the known environments; this is a little bit
more rigid, but also makes it slightly less likely to accidentally
misuse the tool.

[1] https://docs.python.org/3/library/argparse.html