Set up node project
Set up a node project in a subdirectory of the functional-test repo. We
decided we don't want to create a new repo, but in the meantime we don't
want to delete the Python functional test suite. Make it a subdirectory
instead that will eventually overwrite the Python tests.

PER-9318: Write healthcheck test (webclient functional test)