Don't run Python test if JS tests were changed
Since we're creating JS tests in the same repo as Python tests, every
time we open a PR for these tests it will run the Python functional
tests as part of the CI process. We probably want to keep running these
for Dependabots and any bugfixes that need to be done for the Python
test suite, but we don't need to run these every time we update the JS
tests.