Create account only in Sign Up test suite
Instead of creating an account to use throughout the whole suite of
functional tests, we're instead planning on using predefined accounts on
each environment that can be shared by test suites. As a result, only
the Sign Up test suite needs to actually create an account, and cleanup
can be done in its `after` block instead of after running all other
tests. This cleanup hits the API directly instead of testing it through
the web-app.