Update login command to use Cypress session
For upload tests, we want to share session data instead of logging in
each time. Wrap our custom `login` command in a session so that session
data can be reused between tests, and update the signup test suite to
reflect how this command was changed.

PER-9320: Write upload test suite