Remove Javascript Functional Test
Our experiment with a full end-to-end test turned out to be a bit of a
failure. Remove the cypress test suite from the functional test repo so
we don't have to maintain node dependencies anymore.

There is still potential value in doing end-to-end tests, but if we want
to do that one day, we should figure out another way of doing it or
equip the Permanent front end and back end with more capability to be
tested by libraries like cypress.

Remove the /js directory and the eslint action for pull requests. Also
remove the logic that skips running python tests if JS tests were
changed.