Stop failing because of archivematica access copies
Presently, the functional tests are sometimes failing because
Archivematica creates an "access copy" (which is really just the
original) in an unknown format. This happens because Archivematica isn't
presently configured to generate proper access copies of Office files.
For the moment, we should ignore these files, as the don't represent an
actual failure.